+++ lib/dialogs/marker_dialog.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:image_picker/image_picker.dart';
 
-import 'package:helloworld/picture_carousel.dart';
+import 'package:helloworld/widgets/picture_carousel.dart';
 import 'package:helloworld/data/marker.dart';
 
 class MarkerDialog extends StatefulWidget {
